aria-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
httpx>=0.28.0
boto3>=1.35.0
strands>=0.1.0
neo4j>=5.27.0
pymongo>=4.10.0
ddtrace>=2.18.0
copilotkit>=0.1.39
